made latest code compile with ROS2

+++ cartographer_ros/cartographer_ros/node_main.cc
@@ -66,7 +66,7 @@ std::tuple<NodeOptions, TrajectoryOptions> LoadOptions() {
 
 void Run() {
   const auto options = LoadOptions();
-  constexpr double kTfBufferCacheTimeInNs = 1e15; // 1 million seconds
+  constexpr int64_t kTfBufferCacheTimeInNs = 1e15; // 1 million seconds
   tf2_ros::Buffer tf_buffer{::tf2::Duration(kTfBufferCacheTimeInNs)};
   tf2_ros::TransformListener tf(tf_buffer);
   NodeOptions node_options;
